refactor: config jpa by java class
Config jpa by java class. Set hibernate hbm2ddl.auto to validate.
Edit IDs from long to integer and for Truck to string.

+++ logiweb/src/main/java/com/logiweb/avaji/entities/models/CountryMap.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Map;
 
 public class CountryMap {
-    private Map<Long, List<Road>> countryMap;
+    private Map<Integer, List<Road>> countryMap;
 
     public CountryMap() {
