+++ logiweb/src/main/java/com/logiweb/avaji/services/CargoService.java
@@ -15,7 +15,7 @@ public interface CargoService {
      * @param orderId
      * @return list or cargo of specified order
      */
-    List<Cargo> readCargoByOrderId(Long orderId);
+    List<Cargo> readCargoByOrderId(Integer orderId);
 
     /**
      * The <code>createCargo</code> create and check the valid of cargo entity.
